feat: 종료 경고 모달 헤더로 이동
