+++ src/components/TextBubble/TextBubbleService.tsx
@@ -1,7 +1,7 @@
-"use client";
+'use client';
 
-import { useState } from "react";
-import TranslateIcon from "@/assets/icon/generate.svg?react";
+import { useState } from 'react';
+import TranslateIcon from '@/assets/icon/generate.svg?react';
 
 const MOCK_ORIGINAL = `지금 <span class="highlight">결합할인</span>에 대해 알아보고 계시군요
 <span class="highlight">단어</span> 여러개
@@ -25,20 +25,9 @@ export function TextBubbleService() {
         }
       `}</style>
 
-      <div className="flex justify-start my-4">
+      <div className="my-4 flex justify-start">
         {/* 서비스 말풍선 */}
-        <div
-          className="
-            max-w-[240px]
-            px-[10px] py-[10px]
-            bg-white
-            body2
-            rounded-[2px_12px_12px_12px]
-            shadow-[0px_2px_4px_rgba(0,0,0,0.1)]
-            whitespace-pre-wrap
-            break-words
-          "
-        >
+        <div className="body2 max-w-[240px] rounded-[2px_12px_12px_12px] bg-white px-[10px] py-[10px] break-words whitespace-pre-wrap shadow-[0px_2px_4px_rgba(0,0,0,0.1)]">
           <div
             className="whitespace-pre-wrap"
             dangerouslySetInnerHTML={{ __html: MOCK_ORIGINAL }}
@@ -46,9 +35,9 @@ export function TextBubbleService() {
 
           {showSimplified && (
             <>
-              <div className="border-t border-gray-300 my-3" />
+              <div className="my-3 border-t border-gray-300" />
               <div
-                className="text-gray-400 whitespace-pre-wrap"
+                className="whitespace-pre-wrap text-gray-400"
                 dangerouslySetInnerHTML={{ __html: MOCK_SIMPLIFIED }}
               />
             </>
@@ -58,21 +47,11 @@ export function TextBubbleService() {
         {/* 번역 버튼 */}
         <button
           onClick={() => setShowSimplified((prev) => !prev)}
-          className="
-            ml-[10px]
-            mt-[10px]
-            w-[20px] h-[20px]
-            flex items-center justify-center
-            rounded-full
-            bg-gray-300
-            cursor-pointer
-            flex-shrink-0
-          "
+          className="mt-[10px] ml-[10px] flex h-[20px] w-[20px] flex-shrink-0 cursor-pointer items-center justify-center rounded-full bg-gray-300"
         >
-          <TranslateIcon className="w-[10px] h-[10px]" />
+          <TranslateIcon className="h-[10px] w-[10px]" />
         </button>
       </div>
     </>
   );
 }
-
